+++ processor/processor.go
@@ -56,8 +56,6 @@ import (
 	. "github.com/rudderlabs/rudder-server/utils/tx" //nolint:staticcheck
 	reportingtypes "github.com/rudderlabs/rudder-server/utils/types"
 	"github.com/rudderlabs/rudder-server/utils/workerpool"
-	wtrans "github.com/rudderlabs/rudder-server/warehouse/transformer"
-	whutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 )
 
 // Custom type definitions for deeply nested map
@@ -96,18 +94,12 @@ type trackedUsersReporter interface {
 	GenerateReportsFromJobs(jobs []*jobsdb.JobT, sourceIdFilter map[string]bool) []*trackedusers.UsersReport
 }
 
-type warehouseTransformer interface {
-	transformer.DestinationClient
-	CompareAndLog(events []types.TransformerEvent, pResponse, wResponse types.Response)
-}
-
 // Handle is a handle to the processor module
 type Handle struct {
-	conf                 *config.Config
-	tracer               *tracing.Tracer
-	backendConfig        backendconfig.BackendConfig
-	warehouseTransformer warehouseTransformer
-	transformerClients   transformer.TransformerClients
+	conf               *config.Config
+	tracer             *tracing.Tracer
+	backendConfig      backendconfig.BackendConfig
+	transformerClients transformer.TransformerClients
 
 	gatewayDB                  jobsdb.JobsDB
 	routerDB                   jobsdb.JobsDB
@@ -181,7 +173,6 @@ type Handle struct {
 		eventAuditEnabled                         map[string]bool
 		credentialsMap                            map[string][]types.Credential
 		nonEventStreamSources                     map[string]bool
-		enableWarehouseTransformations            config.ValueLoader[bool]
 		enableUpdatedEventNameReporting           config.ValueLoader[bool]
 		enableConcurrentStore                     config.ValueLoader[bool]
 		userTransformationMirroringSanitySampling config.ValueLoader[float64]
@@ -570,8 +561,6 @@ func (proc *Handle) Setup(
 		})
 	}
 
-	proc.warehouseTransformer = wtrans.New(proc.conf, proc.logger, proc.statsFactory)
-
 	if proc.config.enableDedup {
 		var err error
 		proc.dedup, err = dedup.New(proc.conf, proc.statsFactory)
@@ -782,7 +771,6 @@ func (proc *Handle) loadReloadableConfig(defaultPayloadLimit int64, defaultMaxEv
 	proc.config.archivalEnabled = proc.conf.GetReloadableBoolVar(true, "archival.Enabled")
 	// Capture event name as a tag in event level stats
 	proc.config.captureEventNameStats = proc.conf.GetReloadableBoolVar(false, "Processor.Stats.captureEventName")
-	proc.config.enableWarehouseTransformations = proc.conf.GetReloadableBoolVar(false, "Processor.enableWarehouseTransformations")
 	proc.config.enableUpdatedEventNameReporting = proc.conf.GetReloadableBoolVar(false, "Processor.enableUpdatedEventNameReporting")
 	proc.config.enableConcurrentStore = proc.conf.GetReloadableBoolVar(false, "Processor.enableConcurrentStore")
 	// UserTransformation mirroring settings
@@ -2939,7 +2927,7 @@ func (proc *Handle) userTransformAndFilter(
 					}
 
 					if proc.utSamplingFileManager == nil { // Cannot upload, we should just report the issue with no diff
-						log.Errorn("UserTransform sanity check failed")
+						log.Warnn("UserTransform sanity check failed")
 						return
 					}
 
@@ -2969,7 +2957,7 @@ func (proc *Handle) userTransformAndFilter(
 						return
 					}
 
-					log.Errorn("UserTransform sanity check failed",
+					log.Warnn("UserTransform sanity check failed",
 						logger.NewStringField("diffLocation", diffFile.Location),
 						logger.NewStringField("diffObjectName", diffFile.ObjectName),
 						logger.NewStringField("clientEventsLocation", clientEventsFile.Location),
@@ -3174,7 +3162,6 @@ func (proc *Handle) destTransform(
 			proc.logger.Debug("Dest Transform input size", len(data.eventsToTransform))
 			s := time.Now()
 			response = proc.transformerClients.Destination().Transform(ctx, data.eventsToTransform)
-			proc.handleWarehouseTransformations(ctx, data.eventsToTransform, response, data.commonMetaData)
 
 			destTransformationStat := proc.newDestinationTransformationStat(sourceID, workspaceID, transformAt, destination)
 			destTransformationStat.transformTime.Since(s)
@@ -3335,26 +3322,6 @@ func (proc *Handle) destTransform(
 		routerDestIDs:   routerDestIDs,
 		droppedJobs:     data.droppedJobs,
 	}
-}
-
-func (proc *Handle) handleWarehouseTransformations(
-	ctx context.Context,
-	eventsToTransform []types.TransformerEvent,
-	pResponse types.Response,
-	commonMetaData *types.Metadata,
-) {
-	if len(eventsToTransform) == 0 {
-		return
-	}
-	if _, ok := whutils.WarehouseDestinationMap[commonMetaData.DestinationType]; !ok {
-		return
-	}
-	if !proc.config.enableWarehouseTransformations.Load() {
-		return
-	}
-
-	wResponse := proc.warehouseTransformer.Transform(ctx, eventsToTransform)
-	proc.warehouseTransformer.CompareAndLog(eventsToTransform, pResponse, wResponse)
 }
 
 func (proc *Handle) saveDroppedJobs(ctx context.Context, droppedJobs []*jobsdb.JobT, tx *Tx) error {