+++ router/handle_lifecycle.go
@@ -31,6 +31,8 @@ import (
 	"github.com/rudderlabs/rudder-server/rruntime"
 	destinationdebugger "github.com/rudderlabs/rudder-server/services/debugger/destination"
 	"github.com/rudderlabs/rudder-server/services/oauth"
+	oauthv2 "github.com/rudderlabs/rudder-server/services/oauth/v2"
+	common "github.com/rudderlabs/rudder-server/services/oauth/v2/common"
 	"github.com/rudderlabs/rudder-server/services/rmetrics"
 	"github.com/rudderlabs/rudder-server/services/rsources"
 	transformerFeaturesService "github.com/rudderlabs/rudder-server/services/transformer"
@@ -53,6 +55,7 @@ func (rt *Handle) Setup(
 	debugger destinationdebugger.DestinationDebugger,
 	throttlerFactory throttler.Factory,
 	pendingEventsRegistry rmetrics.PendingEventsRegistry,
+	account *backendconfig.AccountWithDefinition,
 ) {
 	rt.backendConfig = backendConfig
 	rt.debugger = debugger
@@ -130,7 +133,12 @@ func (rt *Handle) Setup(
 		rt.reloadableConfig.oauthV2ExpirationTimeDiff,
 		rt.transformerFeaturesService,
 	)
-	rt.isOAuthDestination = oauth.IsOAuthDestination(destinationDefinition.Config)
+	destination := oauthv2.DestinationInfo{
+		Config:  destinationDefinition.Config,
+		Account: account,
+	}
+
+	rt.isOAuthDestination, _ = destination.IsOAuthDestination(common.RudderFlowDelivery)
 	rt.oauth = oauth.NewOAuthErrorHandler(backendConfig)
 
 	rt.isBackendConfigInitialized = false